requirements.txt
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.6.3
timm==0.4.12
sklearn 

tqdm
pillow
matplotlib
numpy

intern
umap-learn
